Refactor: Renames files (again)
The reason the filenames are being changed is that
some languages (Erlang in this case) needs to have
the same module name as the file it is contained in,
and a name like 001 is not a valid Erlang module name.
Thus, a 'p' is being added. At least is not as verbose
as 'problem' again.